more work on hw5.

+++ hw5/controllers/default.py
@@ -14,21 +14,24 @@ def boards():
 
 def board() :
     board = db.boards(request.args(0))
-    return {}
+    return {"board_id":request.args(0)}
 
 from gluon.contrib.populate import populate
 
 def populatedb() :
     populate(db.auth_user,2)
-    populate(db.boards,6)
-    populate(db.posts,40)
+    populate(db.boards,3)
+    populate(db.posts,12)
     boards = db(db.boards).select(orderby=db.boards.updated_at) 
     redirect(URL('default', 'boards'))
 
 def deletedb() :
     db.posts.truncate()
     db.boards.truncate()
-    db.auth_user.truncate()
+    # db.auth_user.truncate()
+    redirect(URL('default', 'boards'))
+
+def home() :
     redirect(URL('default', 'boards'))
 
 def do_logout() :
@@ -39,6 +42,33 @@ def do_login() :
 
 def do_register() :
     redirect(URL('default', 'user', args=['register']))
+
+@auth.requires_signature()
+def add_post():
+    post = db(db.posts.post_id == request.vars.post_id).select().first()
+    board = db(db.boards.id == request.vars.board).select().first()
+    if post :
+        post.update_record(
+            title=request.vars.title,
+            dtitle=request.vars.dtitle,
+            board=board,
+            body=request.vars.body,
+            dbody=request.vars.dbody,
+            author= request.vars.author,
+            is_draft=json.loads(request.vars.is_draft)
+        )
+        return post #load_post(request.vars.board_id);
+
+    db['posts'].insert(
+        title=request.vars.title,
+        dtitle=request.vars.dtitle,
+        board=board,
+        body=request.vars.body,
+        dbody=request.vars.dbody,
+        author= request.vars.author,
+        is_draft=json.loads(request.vars.is_draft)
+    )
+    return post #load_post(request.vars.board_id);
 
 @auth.requires_signature()
 def add_board():
@@ -67,10 +97,16 @@ def add_board():
     return load_board(request.vars.board_id);
 
 @auth.requires_signature()
-def remove_board():
-    print request.vars.id
-    return db(db.boards.board_id == request.vars.board_id).delete()
+def remove_post():
+    return db(db.posts.post_id == request.vars.post_id).delete()
 
+@auth.requires_signature()
+def remove_board():
+    if not auth.user_id : return False
+
+    if(auth.user_id != db(db.boards.board_id == request.vars.board_id).select.first().creator) :
+        return False
+    return db(db.boards.board_id == request.vars.board_id).delete()
 
 
 @auth.requires_signature()
@@ -106,9 +142,41 @@ def load_board(board_id) :
                              board_dict= board_dict
                          ))
 
+def load_posts():
+    ret =load_posts_internal(request.vars.board_id);
+    return ret
+
+def load_posts_internal(board_id) :
+    """Loads One Board and all the posts for that board"""
+    board = db(db.boards.id == board_id).select().first()
+    posts = db(db.posts.board == board_id).select(orderby=~db.posts.created_at)
+
+    my_posts = []
+    if auth.user and auth.user_id :
+        my_posts = [b for b in posts if b.author == auth.user_id]
+
+    board_dict = {
+                  "board_id" : board.board_id,
+                  "creator" : board.creator,
+                  "title"    : board.title,
+                  "dtitle"    : board.dtitle,
+                  "num_posts" : db(db.posts.board == board).count(),
+                  "description" : board.description,
+                  "ddescription" : board.ddescription,
+                  "created_at" : board.created_at.isoformat(),
+                  "is_draft" : board.is_draft,
+                  "uri" : URL('default', 'board', args=[str(board.id)])
+                  }
+
+    return response.json(dict(
+                             board_dict= board_dict,
+                             post_list = posts,
+                             user_id = (auth.user_id if auth and auth.user else -1)
+                         ))
+
 def load_boards():
     """Loads all boards for the user."""
-    boards = db(db.boards).select(orderby=~db.boards.created_at) 
+    boards = db(db.boards).select(orderby=~db.boards.updated_at) 
     my_boards = []
     if auth.user and auth.user_id :
         my_boards = [b for b in boards if b.creator == auth.user_id]
